backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
retell-sdk==1.3.3
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==23.2.1
